Updated acceptance test of voiceover admin to include removing of role.
