+++ core/templates/pages/exploration-editor-page/settings-tab/templates/remove-role-confirmation-modal.component.html
@@ -1,5 +1,5 @@
 <div class="modal-header">
-  <h3 class="oppia-remove-role-confirmation-modal-header">Are you sure?</h3>
+  <h3 class="oppia-remove-role-confirmation-modal-header e2e-test-model-header">Are you sure?</h3>
 </div>
 
 <div class="modal-body oppia-remove-role-confirmation-modal-body">
@@ -7,8 +7,8 @@
 </div>
 
 <div class="modal-footer">
-  <button class="btn btn-secondary" (click)="cancel()">Cancel</button>
-  <button class="btn btn-danger e2e-test-really-delete-exploration-button" (click)="confirm()">Remove user</button>
+  <button class="btn btn-secondary e2e-test-cancel-action-button" (click)="cancel()">Cancel</button>
+  <button class="btn btn-danger e2e-test-really-delete-exploration-button e2e-test-confirm-action-button" (click)="confirm()">Remove user</button>
 </div>
 
 <style>
